requirements.txt
google-generativeai
dotenv
termcolor